# back/services/gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "陪玩平台 Gateway 服务 API 文档",
        "title": "陪玩平台 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/",
    "paths": {
...
        "/api/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户或指定用户的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取用户信息",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "用户ID（可选，不传则获取当前用户）",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "用户UID（可选）",
                        "name": "uid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "手机号（可选）",
                        "name": "phone",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.GetUserResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新当前登录用户的个人信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "更新用户信息",
                "parameters": [
                    {
                        "description": "更新用户信息请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.UpdateUserResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
        "/api/user/avatar": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "上传用户头像，支持 jpg、png、gif、webp 格式，最大 5MB",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "上传头像",
                "parameters": [
                    {
                        "type": "file",
                        "description": "头像文件",
                        "name": "avatar",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.UploadAvatarResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.BaseResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.GetUserResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.UserInfo"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "description": "头像URL",
                    "type": "string"
                },
                "bio": {
                    "description": "个人简介",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "description": "用户角色",
                    "type": "integer"
                }
            }
        },
        "types.UpdateUserResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.UserInfo"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "types.UploadAvatarData": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "description": "头像URL",
                    "type": "string"
                }
            }
        },
        "types.UploadAvatarResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.UploadAvatarData"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.UserInfo": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "description": "头像URL",
                    "type": "string"
                },
                "balance": {
                    "description": "帅币可用余额",
                    "type": "integer"
                },
                "bio": {
                    "description": "个人简介",
                    "type": "string"
                },
                "followerCount": {
                    "description": "粉丝数",
                    "type": "integer"
                },
                "followingCount": {
                    "description": "关注数",
                    "type": "integer"
                },
                "frozenBalance": {
                    "description": "冻结帅币余额（预留）",
                    "type": "integer"
                },
                "id": {
                    "description": "用户ID",
                    "type": "integer"
                },
                "nickname": {
                    "description": "用户昵称",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号码",
                    "type": "string"
                },
                "role": {
                    "description": "用户角色：1=老板, 2=陪玩, 3=管理员",
                    "type": "integer"
                },
                "uid": {
                    "description": "用户唯一标识",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Token，格式: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}